Remove "atomic" package in favour of sync.Mutex
Reseting the counter with the "atomic" package is a little shaky and
probably doesn't work properly. Replace with sync.Mutex

Also simplify the case when we want to send more requests than what is
in the queue by using a non-blocking read.

+++ abios_sdk.go
@@ -111,8 +111,8 @@ func New(username, password string) *client {
 // SetRate sets the outgoing rate to "second" requests per second and "minute" requests
 // per minte. A value less than or equal to 0 means previous
 // value is kept. Default values are (5, 300)
-func (a *client) SetRate(second, minute int) {
-	a.handler.setRate(int32(second), int32(minute))
+func (a *client) SetRate(second, minute uint) {
+	a.handler.setRate(second, minute)
 }
 
 // authenticate queries the /oauth/access_token endpoint with the given credentials and
